Update to amzn2-virtualbox-2.0.20221004.0
X11 libraries are now installed when configuring the VirtualBox guest additions
so that the setup script does not needlessly rebuild the vboxdrv kernel module
at every reboot.  If these libraries are not present, the VBoxClient utility
fails to load, and the vboxadd script will issue the following error:

/opt/VBoxGuestAdditions-6.1.40/bin/VBoxClient: error while loading shared libraries: libX11.so.6: cannot open shared object file: No such file or directory

Because of this error, the vboxadd script assumes that the kernel modules must
be rebuilt at boot time, even if the same kernel and VirtualBox versions are
already present.